UI and logic fixes
